@@ -4,17 +4,16 @@ group = 'com.github.kevinvane'
 
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.2"
+    compileSdkVersion 30
 
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 5
-        versionName "1.5"
+        targetSdkVersion 30
+        versionCode 20
+        versionName "2.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
     }
 
@@ -39,9 +38,9 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:appcompat-v7:28.+'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
 
-    implementation 'com.yanzhenjie:permission:2.0.0-rc5'//不要升级，新版本依赖androidx
+    implementation 'com.yanzhenjie:permission:2.0.3'
     implementation 'com.liulishuo.okdownload:okdownload:1.0.7'
     implementation 'com.liulishuo.okdownload:okhttp:1.0.7'
     implementation 'com.squareup.okhttp3:okhttp:3.12.0' //拦截依赖
